+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Java] Display hints when the glue is not efficiently configured ([#3151](https://github.com/cucumber/cucumber-jvm/pull/3151), Julien Kronegg)
 - [JUnit Platform] Use `ParallelHierarchicalTestExecutorServiceFactory` ([#3105](https://github.com/cucumber/cucumber-jvm/pull/3105))
 - [Core] Support both DocString and DataTable arguments on steps ([#3213](https://github.com/cucumber/cucumber-jvm/pull/3213))
+- [Core] Support `cucumber.execution.threads` property when executing from CLI ([#3183](https://github.com/cucumber/cucumber-jvm/pull/3183))
 
 ### Fixed
 - [JUnit Platform Engine] Accept partial matches with `cucumber.filter.name` and align the behavior with JUnit 4 and CLI ([#3174](https://github.com/cucumber/cucumber-jvm/pull/3174))
@@ -48,6 +49,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [OpenEJB] Removed `cucumber-openejb` in favor of `cucumber-jakarta-openejb` ([#3189](https://github.com/cucumber/cucumber-jvm/pull/3189))
 - [OpenEJB] Removed `cucumber-cdi2` in favor of `cucumber-jakarta-cdi` ([#3192](https://github.com/cucumber/cucumber-jvm/pull/3192))
 - [OpenEJB] Removed `cucumber-deltaspike` without replacement ([#3193](https://github.com/cucumber/cucumber-jvm/pull/3193))
+
+## [7.34.6] - 2026-07-25
+### Fixed
+- [JUnit Platform Engine] Do not fail discovery on an empty rerun file with JUnit Platform 6.1.x ([#3216](https://github.com/cucumber/cucumber-jvm/pull/3216))
+
+## [7.34.5] - 2026-07-23
+### Fixed
+- [Core] Update dependency io.cucumber:messages-ndjson to v0.3.5
+- [Core] Update dependency com.fasterxml.jackson:jackson-bom to v2.22.1
 
 ## [7.34.4] - 2026-06-26
 ### Fixed
@@ -637,7 +647,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Weld] Removed `cucumber-weld` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 - [Needle] Removed `cucumber-needled` in favour of `cucumber-jakarta-cdi` or `cucumber-cdi2`. ([#2276](https://github.com/cucumber/cucumber-jvm/issues/2276) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.4...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.6...HEAD
+[7.34.6]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.5...v7.34.6
+[7.34.5]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.4...v7.34.5
 [7.34.4]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.3...v7.34.4
 [7.34.3]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.2...v7.34.3
 [7.34.2]: https://github.com/cucumber/cucumber-jvm/compare/v7.34.1...v7.34.2